+++ alvr_server/ResampleUtils.h
@@ -59,7 +59,7 @@ public:
 		* ensuring that the output buffer will contain at least all the
 		* converted input samples */
 		max_dst_nb_samples = dst_nb_samples =
-			av_rescale_rnd(default_src_nb_samples, dst_rate, src_rate, AV_ROUND_UP);
+			(int) av_rescale_rnd(default_src_nb_samples, dst_rate, src_rate, AV_ROUND_UP);
 		/* buffer is going to be directly written to a rawaudio file, no alignment */
 		dst_nb_channels = av_get_channel_layout_nb_channels(dst_ch_layout);
 		ret = av_samples_alloc_array_and_samples(&dst_data, &dst_linesize, dst_nb_channels,
@@ -76,7 +76,7 @@ public:
 		int ret;
 
 		/* compute destination number of samples */
-		dst_nb_samples = av_rescale_rnd(swr_get_delay(swr_ctx, src_rate) +
+		dst_nb_samples = (int) av_rescale_rnd(swr_get_delay(swr_ctx, src_rate) +
 			src_nb_samples, dst_rate, src_rate, AV_ROUND_UP);
 		if (dst_nb_samples > max_dst_nb_samples) {
 			av_free(dst_data[0]);
